some refactoring to allow adding patch tests in protego
