+++ .vscode/tasks.json
@@ -41,6 +41,19 @@
       }
     },
     {
+      "label": "make_debug_protego",
+      "type": "shell",
+      "command": "bash",
+      "args": [
+        "-c",
+        ". $(dirname $(which spack))/../share/spack/setup-env.sh && spack env activate aperi-mech && ./do_configure --build-type Debug -p && cd protego-mech/build/Debug && make -j 4"
+      ],
+      "group": {
+        "kind": "build",
+        "isDefault": false
+      }
+    },
+    {
       "label": "make_debug_gpu",
       "type": "shell",
       "command": "bash",
