+++ cmake/testing.cmake
@@ -7,6 +7,10 @@ file(GLOB TEST_SOURCES "test/unit_tests/*.cpp")
 if(USE_PROTEGO_MECH)
     file(GLOB PROTEGO_MECH_TEST_SOURCES "protego-mech/test/unit_tests/*.cpp")
     list(APPEND TEST_SOURCES ${PROTEGO_MECH_TEST_SOURCES})
+    # Remove some tests that wont work with protego-mech. TODO(jake): Some of these tests can be fixed to work with protego-mech
+    list(REMOVE_ITEM TEST_SOURCES "${CMAKE_SOURCE_DIR}/test/unit_tests/ValueFromGeneralizedFieldProcessorTest.cpp")
+    list(REMOVE_ITEM TEST_SOURCES "${CMAKE_SOURCE_DIR}/test/unit_tests/FunctionValueStorageProcessorTest.cpp")
+    list(REMOVE_ITEM TEST_SOURCES "${CMAKE_SOURCE_DIR}/test/unit_tests/CompadreApproximationFunctionTest.cpp")
 endif()
 
 # Add an executable for the unit tests
@@ -22,6 +26,7 @@ target_link_libraries(unit_tests
     aperimech
     GTest::gtest_main
 )
+target_compile_definitions(unit_tests PRIVATE USE_PROTEGO_MECH)
 gtest_discover_tests(unit_tests TIMEOUT 3600 DISCOVERY_TIMEOUT 600)
 
 #--------------------------------------------------
