+++ test/unit_tests/SolverTestFixture.h
@@ -30,16 +30,13 @@ class SolverTest : public ApplicationTest {
     }
 
     double RunSolver() {
+        int procedure_id = 0;
+
         // Create an IO input file object and read the input file
         m_io_input_file = aperi::CreateIoInputFile(m_yaml_data);
 
-        // Create an IO mesh object
-        aperi::IoMeshParameters io_mesh_parameters;  // Default parameters
-        io_mesh_parameters.compose_output = true;
-        m_io_mesh = CreateIoMesh(m_comm, io_mesh_parameters);
-
         // Get parts
-        std::vector<YAML::Node> parts = m_io_input_file->GetParts(0);
+        std::vector<YAML::Node> parts = m_io_input_file->GetParts(procedure_id);
 
         // Get the part names
         std::vector<std::string> part_names;
@@ -53,13 +50,18 @@ class SolverTest : public ApplicationTest {
             }
         }
 
+        // Create an IO mesh object
+        aperi::IoMeshParameters io_mesh_parameters;  // Default parameters
+        io_mesh_parameters.compose_output = true;
+        m_io_mesh = CreateIoMesh(m_comm, io_mesh_parameters);
+
         // Read the mesh
-        m_io_mesh->ReadMesh(m_io_input_file->GetMeshFile(0), part_names);
+        m_io_mesh->ReadMesh(m_io_input_file->GetMeshFile(procedure_id), part_names);
 
         bool uses_generalized_fields = false;
 
         // Loop over parts, create materials, and add parts to force contributions
-        for (auto part : parts) {
+        for (const auto& part : parts) {
             // Create InternalForceContributionParameters
             aperi::InternalForceContributionParameters internal_force_contribution_parameters(part, m_io_input_file, m_io_mesh->GetMeshData());
             m_internal_force_contributions.push_back(aperi::CreateInternalForceContribution(internal_force_contribution_parameters));
@@ -67,7 +69,7 @@ class SolverTest : public ApplicationTest {
         }
 
         // Get field data
-        std::vector<aperi::FieldData> field_data = aperi::GetFieldData(uses_generalized_fields, has_strain_smoothing);
+        std::vector<aperi::FieldData> field_data = aperi::GetFieldData(uses_generalized_fields, has_strain_smoothing, false /* add_debug_fields */);
 
         // Create a mesh labeler
         std::shared_ptr<aperi::MeshLabeler> mesh_labeler = aperi::CreateMeshLabeler();
@@ -86,10 +88,10 @@ class SolverTest : public ApplicationTest {
         }
 
         // Create the field results file
-        m_io_mesh->CreateFieldResultsFile(m_io_input_file->GetOutputFile(0), field_data);
+        m_io_mesh->CreateFieldResultsFile(m_io_input_file->GetOutputFile(procedure_id), field_data);
 
         // Get loads
-        std::vector<YAML::Node> loads = m_io_input_file->GetLoads(0);
+        std::vector<YAML::Node> loads = m_io_input_file->GetLoads(procedure_id);
 
         // Loop over loads and add them to force contributions
         for (auto load : loads) {
@@ -97,11 +99,11 @@ class SolverTest : public ApplicationTest {
         }
 
         // Set initial conditions
-        std::vector<YAML::Node> initial_conditions = m_io_input_file->GetInitialConditions(0);
+        std::vector<YAML::Node> initial_conditions = m_io_input_file->GetInitialConditions(procedure_id);
         aperi::AddInitialConditions(initial_conditions, m_io_mesh->GetMeshData());
 
         // Get boundary conditions
-        std::vector<YAML::Node> boundary_conditions = m_io_input_file->GetBoundaryConditions(0);
+        std::vector<YAML::Node> boundary_conditions = m_io_input_file->GetBoundaryConditions(procedure_id);
 
         // Loop over boundary conditions and add them to the vector of boundary conditions
         for (auto boundary_condition : boundary_conditions) {
@@ -111,10 +113,10 @@ class SolverTest : public ApplicationTest {
         }
 
         // Get the time stepper
-        std::shared_ptr<aperi::TimeStepper> time_stepper = aperi::CreateTimeStepper(m_io_input_file->GetTimeStepper(0));
+        std::shared_ptr<aperi::TimeStepper> time_stepper = aperi::CreateTimeStepper(m_io_input_file->GetTimeStepper(procedure_id));
 
         // Get the output scheduler
-        std::shared_ptr<aperi::Scheduler> output_scheduler = aperi::CreateScheduler(m_io_input_file->GetOutputScheduler(0));
+        std::shared_ptr<aperi::Scheduler> output_scheduler = aperi::CreateScheduler(m_io_input_file->GetOutputScheduler(procedure_id));
 
         // Do preprocessing
         aperi::DoPreprocessing(m_io_mesh, m_internal_force_contributions, m_external_force_contributions, m_boundary_conditions);
@@ -122,8 +124,14 @@ class SolverTest : public ApplicationTest {
         // Create solver
         m_solver = aperi::CreateSolver(m_io_mesh, m_internal_force_contributions, m_external_force_contributions, m_boundary_conditions, time_stepper, output_scheduler);
 
-        //// Run solver
-        return m_solver->Solve();
+        // Run solver
+        double time = m_solver->Solve();
+
+        // Finalize the IO mesh
+        MPI_Barrier(m_comm);
+        m_io_mesh->Finalize();
+
+        return time;
     }
 
     void ResetSolverTest() {
